// migrations/snapshots/eccce1460cf7d1e3e787d2544c944935a0a135acf8aecc6f758b57017cccdc78/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Area: {
      readonly activa: CodecTypes['pg/bool@1']['output'];
      readonly descripcion: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly nombre: CodecTypes['pg/text@1']['output'];
    };
    readonly Atencion: {
      readonly agenteId: CodecTypes['pg/int4@1']['output'];
      readonly casoId: CodecTypes['pg/int4@1']['output'];
      readonly diagnostico: CodecTypes['pg/text@1']['output'];
      readonly fecha: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly solucion: CodecTypes['pg/text@1']['output'];
    };
    readonly Caso: {
      readonly agenteId: CodecTypes['pg/int4@1']['output'] | null;
      readonly categoriaId: CodecTypes['pg/int4@1']['output'];
      readonly descripcion: CodecTypes['pg/text@1']['output'];
      readonly estado: 'PENDIENTE' | 'EN_ANALISIS' | 'EN_ATENCION' | 'EN_VALIDACION' | 'CERRADA';
      readonly fechaAsignacion: CodecTypes['pg/timestamptz-string@1']['output'] | null;
      readonly fechaCierre: CodecTypes['pg/timestamptz-string@1']['output'] | null;
      readonly fechaCreacion: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly prioridad: 'P1' | 'P2' | 'P3';
      readonly solicitanteId: CodecTypes['pg/int4@1']['output'];
      readonly tipo: 'INCIDENTE' | 'SOLICITUD';
      readonly titulo: CodecTypes['pg/text@1']['output'];
    };
    readonly Categoria: {
      readonly activa: CodecTypes['pg/bool@1']['output'];
      readonly areaId: CodecTypes['pg/int4@1']['output'];
      readonly descripcion: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly nombre: CodecTypes['pg/text@1']['output'];
    };
    readonly Historial: {
      readonly casoId: CodecTypes['pg/int4@1']['output'];
      readonly comentario: CodecTypes['pg/text@1']['output'] | null;
      readonly estadoAnterior:
        'PENDIENTE' | 'EN_ANALISIS' | 'EN_ATENCION' | 'EN_VALIDACION' | 'CERRADA' | null;
      readonly estadoNuevo:
        'PENDIENTE' | 'EN_ANALISIS' | 'EN_ATENCION' | 'EN_VALIDACION' | 'CERRADA' | null;
      readonly evento:
        | 'CREACION'
        | 'RECLASIFICACION'
        | 'ASIGNACION'
        | 'CAMBIO_ESTADO'
        | 'ATENCION'
        | 'APROBACION'
        | 'DEVOLUCION';
      readonly fecha: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly usuarioId: CodecTypes['pg/int4@1']['output'];
    };
    readonly Usuario: {
      readonly activo: CodecTypes['pg/bool@1']['output'];
      readonly correo: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly nombre: CodecTypes['pg/text@1']['output'];
      readonly rol: 'SOLICITANTE' | 'AGENTE' | 'VALIDADOR' | 'ADMINISTRADOR';
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Area: {
      readonly activa: CodecTypes['pg/bool@1']['input'];
      readonly descripcion: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly nombre: CodecTypes['pg/text@1']['input'];
    };
    readonly Atencion: {
      readonly agenteId: CodecTypes['pg/int4@1']['input'];
      readonly casoId: CodecTypes['pg/int4@1']['input'];
      readonly diagnostico: CodecTypes['pg/text@1']['input'];
      readonly fecha: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly solucion: CodecTypes['pg/text@1']['input'];
    };
    readonly Caso: {
      readonly agenteId: CodecTypes['pg/int4@1']['input'] | null;
      readonly categoriaId: CodecTypes['pg/int4@1']['input'];
      readonly descripcion: CodecTypes['pg/text@1']['input'];
      readonly estado: 'PENDIENTE' | 'EN_ANALISIS' | 'EN_ATENCION' | 'EN_VALIDACION' | 'CERRADA';
      readonly fechaAsignacion: CodecTypes['pg/timestamptz-string@1']['input'] | null;
      readonly fechaCierre: CodecTypes['pg/timestamptz-string@1']['input'] | null;
      readonly fechaCreacion: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly prioridad: 'P1' | 'P2' | 'P3';
      readonly solicitanteId: CodecTypes['pg/int4@1']['input'];
      readonly tipo: 'INCIDENTE' | 'SOLICITUD';
      readonly titulo: CodecTypes['pg/text@1']['input'];
    };
    readonly Categoria: {
      readonly activa: CodecTypes['pg/bool@1']['input'];
      readonly areaId: CodecTypes['pg/int4@1']['input'];
      readonly descripcion: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly nombre: CodecTypes['pg/text@1']['input'];
    };
    readonly Historial: {
      readonly casoId: CodecTypes['pg/int4@1']['input'];
      readonly comentario: CodecTypes['pg/text@1']['input'] | null;
      readonly estadoAnterior:
        'PENDIENTE' | 'EN_ANALISIS' | 'EN_ATENCION' | 'EN_VALIDACION' | 'CERRADA' | null;
      readonly estadoNuevo:
        'PENDIENTE' | 'EN_ANALISIS' | 'EN_ATENCION' | 'EN_VALIDACION' | 'CERRADA' | null;
      readonly evento:
        | 'CREACION'
        | 'RECLASIFICACION'
        | 'ASIGNACION'
        | 'CAMBIO_ESTADO'
        | 'ATENCION'
        | 'APROBACION'
        | 'DEVOLUCION';
      readonly fecha: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly usuarioId: CodecTypes['pg/int4@1']['input'];
    };
    readonly Usuario: {
      readonly activo: CodecTypes['pg/bool@1']['input'];
      readonly correo: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly nombre: CodecTypes['pg/text@1']['input'];
      readonly rol: 'SOLICITANTE' | 'AGENTE' | 'VALIDADOR' | 'ADMINISTRADOR';
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly area: {
      readonly activa: CodecTypes['pg/bool@1']['output'];
      readonly descripcion: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly nombre: CodecTypes['pg/text@1']['output'];
    };
    readonly atencion: {
      readonly agente_id: CodecTypes['pg/int4@1']['output'];
      readonly caso_id: CodecTypes['pg/int4@1']['output'];
      readonly diagnostico: CodecTypes['pg/text@1']['output'];
      readonly fecha: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly solucion: CodecTypes['pg/text@1']['output'];
    };
    readonly caso: {
      readonly agente_id: CodecTypes['pg/int4@1']['output'] | null;
      readonly categoria_id: CodecTypes['pg/int4@1']['output'];
      readonly descripcion: CodecTypes['pg/text@1']['output'];
      readonly estado: 'PENDIENTE' | 'EN_ANALISIS' | 'EN_ATENCION' | 'EN_VALIDACION' | 'CERRADA';
      readonly fecha_asignacion: CodecTypes['pg/timestamptz-string@1']['output'] | null;
      readonly fecha_cierre: CodecTypes['pg/timestamptz-string@1']['output'] | null;
      readonly fecha_creacion: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly prioridad: 'P1' | 'P2' | 'P3';
      readonly tipo: 'INCIDENTE' | 'SOLICITUD';
      readonly titulo: CodecTypes['pg/text@1']['output'];
      readonly usuario_id: CodecTypes['pg/int4@1']['output'];
    };
    readonly categoria: {
      readonly activa: CodecTypes['pg/bool@1']['output'];
      readonly area_id: CodecTypes['pg/int4@1']['output'];
      readonly descripcion: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly nombre: CodecTypes['pg/text@1']['output'];
    };
    readonly historial: {
      readonly caso_id: CodecTypes['pg/int4@1']['output'];
      readonly comentario: CodecTypes['pg/text@1']['output'] | null;
      readonly estado_anterior:
        'PENDIENTE' | 'EN_ANALISIS' | 'EN_ATENCION' | 'EN_VALIDACION' | 'CERRADA' | null;
      readonly estado_nuevo:
        'PENDIENTE' | 'EN_ANALISIS' | 'EN_ATENCION' | 'EN_VALIDACION' | 'CERRADA' | null;
      readonly evento:
        | 'CREACION'
        | 'RECLASIFICACION'
        | 'ASIGNACION'
        | 'CAMBIO_ESTADO'
        | 'ATENCION'
        | 'APROBACION'
        | 'DEVOLUCION';
      readonly fecha: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly usuario_id: CodecTypes['pg/int4@1']['output'];
    };
    readonly usuario: {
      readonly activo: CodecTypes['pg/bool@1']['output'];
      readonly correo: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly nombre: CodecTypes['pg/text@1']['output'];
      readonly rol: 'SOLICITANTE' | 'AGENTE' | 'VALIDADOR' | 'ADMINISTRADOR';
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;